+++ src/datasources/results/defaultQueries.ts
@@ -33,7 +33,7 @@ export const defaultStepsQuery: Omit<QuerySteps, 'refId'> = {
   ] as StepsProperties[],
   orderBy: "STARTED_AT",
   descending: false,
-  recordCount: 10_000,
+  recordCount: 1000,
   useTimeRange: false,
   useTimeRangeFor: undefined,
   resultsQuery: '',